Add error decoders
Change from `Http.Error` to `Http.expectStringResponse`. I probably wouldn't use this and stick with `Http.Error`. We'd have server-side validation, but handle the form errors with Elm (probably @rtfeldman's style).

It's nice to have this as an option though. Especially if it can easily be extended for larger forms.